# REF: use consistent pattern in tslibs.vectorized