@jreback 2020-08-07
> DO NOT MERGE - Performance takes a hit

I think you can do a draft PR to indicate this (of course comments ok too)